+++ index.html
@@ -9,10 +9,12 @@
         <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
         <link href="https://fonts.googleapis.com/css2?family=Inter&display=swap" rel="stylesheet">
         <link rel="stylesheet" href="index.css">
+        <script type="module" src="index.js"></script>
     </head>
     <body>
         <main>
             <header id="header">
+              <div class="wrapper">
                 <h3>Awesome Books</h3>
                 <nav class="nav">
                 <ul class="listStyle">
@@ -21,6 +23,8 @@
                     <li><a href="#contact">Contact</a></li>
                 </ul>
                 </nav>
+              </div>
+              <div id='dateTime'></div>
             </header>
 
             <section id="list" class="books-manager section">
@@ -57,6 +61,5 @@
               <p>Copyright Steven&Nadya 2022</p>
             </footer>
         </main>
-        <script type="module" src="index.js"></script>
     </body>
 </html>
